Connect Plan Route to GUI

+++ src/main/java/org/example/PT/Route.java
@@ -39,7 +39,7 @@ public class Route {
     }
 
 
-    public void calculateRoute(int parcelNumber, int idAddressA,int  idAddressB) {
+    public void calculateRoute(long parcelNumber, int idAddressA,int  idAddressB) {
         setDistanceABC();
         setDistance(idAddressA, idAddressB);
         int distance;
